pyproject.toml
[project]
name = "mrvn"
version = "0.1.0"
description = "A test implementation of an LLM harness"
authors = [{ name = "monkut" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.14,<3.15"
classifiers = [
    "Intended Audience :: Developers",
    "Private :: Do Not Upload",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "django>=5.2.8,<5.3",
    "psycopg[binary]",
    "pydantic<3.0",
    "whitenoise<7.0",
    # Messaging integrations
    "python-telegram-bot>=21.0",
    "slack-bolt>=1.18.0",
    # LLM providers
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "ollama>=0.4.0",
    "openai>=1.50.0",  # For vLLM OpenAI-compatible API
    # Production server
    "gunicorn>=23.0.0",
    "httpx>=0.27.0",  # HTTP client for validations
]

[project.urls]
Repository = "https://github.com/monkut/marvin-manager"

[dependency-groups]
dev = [
    "pyright>=1.1.396",
    "ruff>=0.9.10",
    "unittest-xml-reporting>=3.2.0",
    "poethepoet>=0.34.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["mrvn/*"]  # package names should match these glob patterns (["*"] by default)
exclude = []  # exclude packages matching these glob patterns (empty by default)

# Task runner
[tool.poe.tasks]
check = "uv run ruff check"
typecheck = "uv run pyright"


[tool.poe.tasks.test]
# run the test command in a shell
shell = "cd mrvn && uv run python manage.py test --debug-mode"

[tool.pyright]
include = ["mrvn", "notebooks"]
exclude = ["**/.venv", "tmp/", "**/node_modules/", "**/__pycache__", "**/*.pyc", "**/tests/", "**/migrations/"]
typeCheckingMode = "basic"
pythonVersion = "3.14"

[tool.ruff]
line-length = 120
indent-width = 4
target-version ='py314'
exclude = [".git", ".venv", "**/node_modules/", "**/.volta", "jupyter/*", "**/migrations/", "**/manage.py"]
fix = true
respect-gitignore = true
unsafe-fixes = false

[tool.ruff.lint]
select = ["ALL"]
extend-per-file-ignores = { "**/__init__.py" = ["I", "F403"] }
extend-safe-fixes = [
    "D200",  # unnecessary-multiline-docstring
    "ANN204"  # missing-return-type-special-method (__init__)
]
ignore = [
    "A005",
    "ANN002",
    "ANN003",
    "ANN201",
    "ARG002",
    "ARG003",
    "C401",  # rewrite as a set comprehension
    "COM812",
    "COM819",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D203",
    "D205",
    "D206",
    "D213",
    "D300",
    "D400",
    "D401",
    "D415",
    "DJ008",  # Model does not define `__str__` method
    "DTZ005",
    "E111",
    "E114",
    "E117",
    "EM101",
    "EM102",
    "ERA001",
    "EXE002",  # file is executable but no shebang is present
    "FBT",
    "FIX002",
    "G004",
    "INP001",
    "ISC001",
    "ISC002",
    "PERF401",  # Use `list.extend` to create a transformed list
    "PLR0913",
    "PT009",
    "PT011",
    "PT027",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "RET504",
    "RUF",
    "S101",
    "SIM108",  # if-else-block-instead-of-if-exp
    "SIM118",  # Use `key in dict` instead of `key in dict.keys()`
    "SLF001",
    "TC001",
    "TC002",
    "TD002",  # missing-todo-author
    "TD003",
    "TID252",
    "TRY003",
]

exclude = ["target", "tmp", "data", "notebooks", "**/migrations/"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds this value.
max-complexity = 13

[tool.ruff.lint.isort]
combine-as-imports = true
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
split-on-trailing-comma = true

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"